requirements.txt
asgiref==3.2.10
Django==3.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
